requirements.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.1.11
django-crispy-forms==1.11.2
django-cuser==2017.3.16
django-debug-toolbar==3.2.1
django-easy-audit==1.3.0
django-extensions==3.1.3
django-filter==2.4.0
django-heroku==0.3.1
django-login-required-middleware==0.6.1
django-smart-selects==1.5.9
gunicorn==20.1.0
psycopg2==2.8.6
pytz==2021.1
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
whitenoise==5.2.0
